Decompress and rename contents of CDC archives
Renaming is necessary because the contents of each archive follow a
different naming convention. This makes it difficult to easily
process the files in a loop.

Files are named using the convention [year].[num].zip where [num] is
the order of the file in the archive. I assume archives can contain
multiple files.